+++ targets/pic_16bit/mikroe/dspic/src/uart/implementation_1/hal_ll_uart.c
@@ -125,7 +125,7 @@ typedef struct
     hal_ll_uart_data_bits_t data_bit;
     bool alternate_pins;
     uint8_t hal_ll_pps_module_index;
-    uint16_t timeout_polling_write;
+    uint32_t timeout_polling_write;
 } hal_ll_uart_hw_specifics_map_t;
 
 /*!< @brief UART hw specific module values */
@@ -563,7 +563,7 @@ hal_ll_err_t hal_ll_uart_set_data_bits( handle_t *handle, hal_ll_uart_data_bits_
     return HAL_LL_UART_SUCCESS;
 }
 
-void hal_ll_uart_set_polling_write_timeout( handle_t *handle, uint16_t timeout ) {
+void hal_ll_uart_set_polling_write_timeout( handle_t *handle, uint32_t timeout ) {
     low_level_handle = hal_ll_uart_get_handle;
     hal_ll_uart_hw_specifics_map_local = hal_ll_get_specifics(hal_ll_uart_get_module_state_address);
 
@@ -649,7 +649,7 @@ void hal_ll_uart_write( handle_t *handle, uint8_t wr_data ) {
 
 void hal_ll_uart_write_polling( handle_t *handle, uint8_t wr_data ) {
     const hal_ll_uart_base_handle_t *reg_addresses = hal_ll_uart_get_base_handle;
-    uint16_t time_counter = hal_ll_uart_hw_specifics_map_local->timeout_polling_write;
+    uint32_t time_counter = hal_ll_uart_hw_specifics_map_local->timeout_polling_write;
 
     while ( !check_reg_bit( reg_addresses->uart_sta_reg_addr, HAL_LL_TRANSMIT_SHIFT_REGISTER_BIT ) ) {
         // Timeout check
