+++ targets/pic_32bit/mikroe/pic32/src/uart/implementation_1/hal_ll_uart.c
@@ -283,7 +283,7 @@ typedef struct {
     hal_ll_uart_stop_bits_t stop_bits;
     uint8_t iec_rx_bit;
     uint8_t iec_tx_bit;
-    uint16_t timeout_polling_write;
+    uint32_t timeout_polling_write;
 } hal_ll_uart_hw_specifics_map_t;
 
 /*!< @brief UART hw specific module values */
@@ -696,7 +696,7 @@ hal_ll_err_t hal_ll_uart_set_stop_bits( handle_t *handle, hal_ll_uart_stop_bits_
     return HAL_LL_UART_SUCCESS;
 }
 
-void hal_ll_uart_set_polling_write_timeout( handle_t *handle, uint16_t timeout ) {
+void hal_ll_uart_set_polling_write_timeout( handle_t *handle, uint32_t timeout ) {
     hal_ll_uart_hw_specifics_map_local = hal_ll_get_specifics(hal_ll_uart_get_module_state_address);
 
     if( hal_ll_uart_hw_specifics_map_local->base != HAL_LL_MODULE_ERROR ) {
@@ -791,7 +791,7 @@ void hal_ll_uart_write( handle_t *handle, uint8_t wr_data ) {
 
 void hal_ll_uart_write_polling( handle_t *handle, uint8_t wr_data ) {
     hal_ll_uart_base_handle_t *hal_ll_hw_reg = hal_ll_uart_get_wr_rd_handle;
-    uint16_t time_counter = hal_ll_uart_hw_specifics_map_local->timeout_polling_write;
+    uint32_t time_counter = hal_ll_uart_hw_specifics_map_local->timeout_polling_write;
 
     while ( !check_reg_bit(&hal_ll_hw_reg->uart_sta_reg_addr, HAL_LL_UART_TRMT_BIT) );
